Restructure mappers each into their own files
