+++ src/emulator/mappers/mod.rs
@@ -1,0 +1,5 @@
+mod nrom;
+mod mmc1;
+
+pub use self::nrom::NROM;
+pub use self::mmc1::MMC1;
